Remove race condition
Previously, there was a race condition in the case
that a user specified multiple stacks at the same
time with the same name. It was possible for one of
them to override the stack name file of the first.
This is problematic b/c it can lead to stacks that
are lost from the perspective of stack_builder (they
cannot be listed or destroyed)

This commit makes the initial creation of the stack
name file an atomic operation to remove this
race condition.